+++ content/lib/pyswitch/clients/kemper/mappings/freeze.py
@@ -1,5 +1,6 @@
 from micropython import const
-from .. import KemperParameterMapping, KemperNRPNMessage, KemperEffectSlot
+from .. import KemperNRPNMessage, KemperEffectSlot
+from ....controller.Client import ClientParameterMapping
 from .. import NRPN_FUNCTION_SET_SINGLE_PARAMETER, NRPN_FUNCTION_REQUEST_SINGLE_PARAMETER, NRPN_FUNCTION_RESPONSE_SINGLE_PARAMETER
 
 from adafruit_midi.control_change import ControlChange
@@ -11,7 +12,7 @@ _NRPN_ADDRESS_PAGE_FREEZE = const(0x7d)
 
 # Freeze for slots
 def MAPPING_FREEZE(slot_id):
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = f"Freeze { str(slot_id) }",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -32,7 +33,7 @@ def MAPPING_FREEZE(slot_id):
 
 # Freeze (global) for all reverb and delay modules (no feedback from kemper!)
 def MAPPING_FREEZE_ALL_GLOBAL():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "Freeze",
         set = ControlChange(
             _CC_FREEZE_ALL_GLOBAL,