+++ content/lib/pyswitch/clients/kemper/mappings/looper.py
@@ -1,5 +1,6 @@
 from micropython import const
-from .. import KemperParameterMapping, KemperNRPNMessage
+from .. import KemperNRPNMessage
+from ....controller.Client import ClientParameterMapping
 from .. import NRPN_FUNCTION_SET_SINGLE_PARAMETER
 
 _NRPN_ADDRESS_PAGE_FREEZE = const(0x7d)
@@ -14,7 +15,7 @@ _NRPN_LOOPER_PARAMETER_ERASE = const(94)
 
 
 def MAPPING_LOOPER_REC_PLAY_OVERDUB():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopRec",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -24,7 +25,7 @@ def MAPPING_LOOPER_REC_PLAY_OVERDUB():
     )
 
 def MAPPING_LOOPER_STOP():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopStop",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -34,7 +35,7 @@ def MAPPING_LOOPER_STOP():
     )
 
 def MAPPING_LOOPER_TRIGGER():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopTrig",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -44,7 +45,7 @@ def MAPPING_LOOPER_TRIGGER():
     )
 
 def MAPPING_LOOPER_REVERSE():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopRev",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -54,7 +55,7 @@ def MAPPING_LOOPER_REVERSE():
     )
 
 def MAPPING_LOOPER_HALF_SPEED():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "Loop1/2",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -64,7 +65,7 @@ def MAPPING_LOOPER_HALF_SPEED():
     )
 
 def MAPPING_LOOPER_CANCEL():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopCanc",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER, 
@@ -74,7 +75,7 @@ def MAPPING_LOOPER_CANCEL():
     )
 
 def MAPPING_LOOPER_ERASE():
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "LoopErase",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER,