+++ content/lib/pyswitch/clients/kemper/mappings/rig.py
@@ -1,5 +1,6 @@
 from micropython import const
-from .. import KemperParameterMapping, KemperNRPNMessage
+from .. import KemperNRPNMessage
+from ....controller.Client import ClientParameterMapping
 from .. import NRPN_FUNCTION_REQUEST_STRING_PARAMETER, NRPN_FUNCTION_RESPONSE_STRING_PARAMETER, NRPN_FUNCTION_SET_SINGLE_PARAMETER, NRPN_FUNCTION_RESPONSE_SINGLE_PARAMETER, NRPN_FUNCTION_REQUEST_SINGLE_PARAMETER, NRPN_ADDRESS_PAGE_STRINGS
 
 _NRPN_ADDRESS_PAGE_RIG_PARAMETERS = const(0x04)
@@ -11,7 +12,7 @@ _NRPN_STRING_PARAMETER_ID_RIG_DATE = const(0x03)
 
 # Rig comment (request only)
 def MAPPING_RIG_COMMENT(): 
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "Rig Comment",
         request = KemperNRPNMessage(               
             NRPN_FUNCTION_REQUEST_STRING_PARAMETER,
@@ -23,12 +24,12 @@ def MAPPING_RIG_COMMENT():
             NRPN_ADDRESS_PAGE_STRINGS,
             _NRPN_STRING_PARAMETER_ID_RIG_COMMENT
         ),
-        type = KemperParameterMapping.PARAMETER_TYPE_STRING
+        type = ClientParameterMapping.PARAMETER_TYPE_STRING
     )
 
 # Rig date (request only)
 def MAPPING_RIG_DATE(): 
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "Rig Date",
         request = KemperNRPNMessage(               
             NRPN_FUNCTION_REQUEST_STRING_PARAMETER, 
@@ -40,12 +41,12 @@ def MAPPING_RIG_DATE():
             NRPN_ADDRESS_PAGE_STRINGS,
             _NRPN_STRING_PARAMETER_ID_RIG_DATE
         ),
-        type = KemperParameterMapping.PARAMETER_TYPE_STRING
+        type = ClientParameterMapping.PARAMETER_TYPE_STRING
     )
 
 # Rig volume
 def MAPPING_RIG_VOLUME(): 
-    return KemperParameterMapping(
+    return ClientParameterMapping(
         name = "Rig Volume",
         set = KemperNRPNMessage(
             NRPN_FUNCTION_SET_SINGLE_PARAMETER,