+++ content/lib/pyswitch/controller/Client.py
@@ -1,17 +1,29 @@
+from math import floor
+from micropython import const
 from ..misc import EventEmitter, PeriodCounter, Updateable, get_option, compare_midi_messages, stringify_midi_message, do_print
-#from ..stats import RuntimeStatistics
+
+from adafruit_midi.control_change import ControlChange
+from adafruit_midi.system_exclusive import SystemExclusive
+from adafruit_midi.program_change import ProgramChange
 
 
 # Midi mapping for a client command. Contains commands to set or request a parameter
 class ClientParameterMapping:
+    
+    # Parameter types (used internally in mappings)
+    PARAMETER_TYPE_NUMERIC = const(0)   # Default, also used for on/off
+    PARAMETER_TYPE_STRING = const(1)
+
     # Takes MIDI messages as argument (ControlChange or SystemExclusive)
-    def __init__(self, name = "", set = None, request = None, response = None, value = None):
+    def __init__(self, name = "", set = None, request = None, response = None, value = None, type = 0):
         self.name = name          # Mapping name (used for debug output only)
         self.set = set            # MIDI Message to set the parameter
         self.request = request    # MIDI Message to request the value
         self.response = response  # Response template MIDI message for parsing the received answer        
         self.value = value        # Value of the parameter (buffer). After receiving an answer, the value 
-                                  # is buffered here.                                  
+                                  # is buffered here.
+        self.type = type          # NUmeric or string
+
 
     def __eq__(self, other):
         if not other:
@@ -59,21 +71,151 @@ class ClientParameterMapping:
             else:
                 return False
             
-    # Must parse the incoming MIDI message and set its value on the mapping.
-    # If the response template does not match, must return False, and
-    # vice versa must return True to notify the listeners of a value change.
-    def parse(self, midi_message):   # pragma: no cover
-        return False    
+    # Parse the incoming MIDI message and set its value on the mapping.
+    # If the response template does not match, returns False, and
+    # vice versa. Returns True to notify the listeners of a value change.
+    def parse(self, midi_message):     
+        result = self.parse_against(midi_message, self.response)
+        if result != None:
+            self.value = result
+            return True
+        
+        return False
+
+    # Parse a message against a response message
+    def parse_against(self, midi_message, response):     
+        # SysEx (NRPN) Messages
+        if isinstance(midi_message, SystemExclusive):
+            if not isinstance(response, SystemExclusive):
+                return None
+                     
+            # Compare manufacturer IDs
+            if midi_message.manufacturer_id != response.manufacturer_id:
+                return None
+            
+            # Check if the message belongs to the mapping. The following have to match:
+            #   2: function code, 
+            #   3: instance ID, 
+            #   4: address page, 
+            #   5: address nunber
+            #
+            # The first two values are ignored (the Kemper MIDI specification implies this would contain the product type
+            # and device ID as for the request, however the device just sends two zeroes)
+            if midi_message.data[2:6] != response.data[2:6]:
+                return None
+            
+            # The values starting from index 6 are the value of the response.
+            if self.type == self.PARAMETER_TYPE_STRING:
+                # Take as string
+                return ''.join(chr(int(c)) for c in list(midi_message.data[6:-1]))
+            else:
+                # Decode 14-bit value to int
+                return midi_message.data[-2] * 128 + midi_message.data[-1]
+
+        # CC Messages
+        elif isinstance(midi_message, ControlChange):
+            if not isinstance(response, ControlChange):
+                return None
+            
+            if midi_message.control == response.control:
+                return midi_message.value
+
+        # PC Messages
+        elif isinstance(midi_message, ProgramChange):
+            if not isinstance(response, ProgramChange):
+                return None
+            
+            return midi_message.patch
+
+        # MIDI Clock
+        #elif isinstance(midi_message, MidiClockMessage):
+        #    self._count_clock += 1
+
+        #    if self._count_clock >= 24:
+        #        self._count_clock = 0
+
+        #    if self._count_clock >= 12:
+        #        mapping.value = 1
+        #    else:
+        #        mapping.value = 0    
+
+        #    return True
+        
+        # MIDI Clock start
+        #elif isinstance(midi_message, Start):
+        #    self._count_clock = 0
+
+        #    mapping.value = 1
+
+        #    return True
+
+        return None
     
-    # Must set the passed value(s) on the SET message(s) of the mapping.
-    def set_value(self, value):      # pragma: no cover
-        pass
+    # Set the passed value(s) on the SET message(s) of the mapping.
+    def set_value(self, value):
+        if isinstance(self.set, list):
+            for i in range(len(self.set)):
+                self.__set_value(self.set[i], value[i])
+        else:
+            self.__set_value(self.set, value)
+
+    def __set_value(self, midi_message, value):
+        if self.type == self.PARAMETER_TYPE_STRING:
+            raise Exception() # Setting strings is not implemented yet
+
+        if isinstance(midi_message, ControlChange):
+            # Set value directly (CC takes int values)            
+            midi_message.value = value
+
+        elif isinstance(midi_message, SystemExclusive):            
+            # Fill up message to appropriate length for the specification
+            data = list(midi_message.data)
+            while len(data) < 8:
+                data.append(0)
+            
+            # Set value as 14 bit
+            data[6] = int(floor(value / 128))
+            data[7] = int(value % 128)
+
+            midi_message.data = bytes(data)
+
+        elif isinstance(midi_message, ProgramChange):
+            # Set patch
+            midi_message.patch = value
 
     # Returns if the mapping has finished receiving a result. Per default,
     # this returns True which is valid for mappings with one response.
     def result_finished(self):
         return True
     
+
+# Parser for two-part messages: The result value will be 128 * value1 + value2, 
+# notified when the second message arrives.
+class ClientTwoPartParameterMapping(ClientParameterMapping):
+
+    def __init__(self, name = "", set = None, request = None, response = None, value = None, type = 0):
+        super().__init__(name = name, set = set, request = request, response = response, value = value, type = type)
+
+        self.__value_1 = None
+    
+    def parse(self, midi_message): 
+        value_1 = self.parse_against(midi_message, self.response[0])
+        if value_1 != None:
+            self.__value_1 = value_1
+            return True
+        
+        value_2 = self.parse_against(midi_message, self.response[1])
+
+        if value_2 != None and self.__value_1 != None:
+            self.value = 128 * self.__value_1 + value_2
+            self.__value_1 = None
+            return True
+        
+        return False
+            
+    def result_finished(self):
+        return (self.__value_1 == None)
+
 
 ############################################################################################################