+++ test/pyswitch/test_client_mapping_parser.py
@@ -20,10 +20,10 @@ with patch.dict(sys.modules, {
     from adafruit_midi.control_change import ControlChange
     from adafruit_midi.program_change import ProgramChange
 
-    from lib.pyswitch.clients.kemper import *
+    from lib.pyswitch.controller.Client import *
 
 
-class TestKemperParameterMappings(unittest.TestCase):
+class TestClientParameterMappingParser(unittest.TestCase):
 
     def test_parse_sysex(self):
         msg_irrelevant_man_id = SystemExclusive(
@@ -80,7 +80,7 @@ class TestKemperParameterMappings(unittest.TestCase):
                 ]
             )            
 
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             response = SystemExclusive(
                 manufacturer_id = [0x00, 0x10, 0x20],
                 data = [0x00, 0x00, 0xd9, 0x01, 0x04, 0xaa, 0x00, 0x00]
@@ -140,12 +140,12 @@ class TestKemperParameterMappings(unittest.TestCase):
                 ] + hex_str + [0]
             )            
         
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             response = SystemExclusive(
                 manufacturer_id = [0x00, 0x10, 0x20],
                 data = [0x00, 0x00, 0xd9, 0x01, 0x04, 0xaa]
             ),
-            type = KemperParameterMapping.PARAMETER_TYPE_STRING       
+            type = ClientParameterMapping.PARAMETER_TYPE_STRING       
         )
 
         # Valid cases
@@ -187,7 +187,7 @@ class TestKemperParameterMappings(unittest.TestCase):
             value = 0
         )           
                 
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             response = ControlChange(
                 control = 1,
                 value = 0
@@ -236,7 +236,7 @@ class TestKemperParameterMappings(unittest.TestCase):
             patch = 0
         )
                 
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             response = ProgramChange(
                 patch = 0
             )   
@@ -264,7 +264,7 @@ class TestKemperParameterMappings(unittest.TestCase):
 
 
     def test_parse_other(self):         
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             response = ProgramChange(
                 patch = 0
             )   
@@ -278,7 +278,7 @@ class TestKemperParameterMappings(unittest.TestCase):
 
 
     def test_set_value_sysex_numeric(self):
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = SystemExclusive(
                 manufacturer_id = [0x00, 0x10, 0x20],
                 data = [0x00, 0x00, 0xd9, 0x01, 0x04, 0xaa]
@@ -321,12 +321,12 @@ class TestKemperParameterMappings(unittest.TestCase):
 
 
     def test_set_value_sysex_string(self):
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = SystemExclusive(
                 manufacturer_id = [0x00, 0x10, 0x20],
                 data = [0x00, 0x00, 0xd9, 0x01, 0x04, 0xaa]
             ),
-            type = KemperParameterMapping.PARAMETER_TYPE_STRING
+            type = ClientParameterMapping.PARAMETER_TYPE_STRING
         )
 
         with self.assertRaises(Exception):
@@ -334,7 +334,7 @@ class TestKemperParameterMappings(unittest.TestCase):
         
 
     def test_set_value_control_change(self):
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = ControlChange(
                 control = 8,
                 value = 0
@@ -354,7 +354,7 @@ class TestKemperParameterMappings(unittest.TestCase):
 
 
     def test_set_value_program_change(self):
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = ProgramChange(
                 patch = 8
             )      
@@ -375,7 +375,7 @@ class TestKemperParameterMappings(unittest.TestCase):
             def __init__(self, value):
                 self.value = value
 
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = FooMessage(9) 
         )
 
@@ -390,7 +390,7 @@ class TestKemperParameterMappings(unittest.TestCase):
 
 
     def test_set_value_list(self):
-        mapping = KemperParameterMapping(
+        mapping = ClientParameterMapping(
             set = [
                 ProgramChange(
                     patch = 8
@@ -441,7 +441,7 @@ class TestKemperParameterMappings(unittest.TestCase):
                 ]
             )            
 
-        mapping = KemperTwoPartParameterMapping(
+        mapping = ClientTwoPartParameterMapping(
             response = [
                 ControlChange(
                     control = 112,